\name{mota-internal}
\title{Internal MOTA function}
\alias{pvalue.mota}
\alias{r2.mota}

\description{Internal MOTA function}

\usage{
pvalue.mota(object, numberOfIterations = 50)
r2.mota(object, ...)

}
\author{Stefan Hengl}

\details{Not to be called by the user}
\keyword{internal}

