\name{mota-internal}
\title{Internal MOTA function}
\alias{pvalue.mota}
\alias{r2.mota}
\alias{mota.progress}
\description{Internal MOTA function}

\usage{
pvalue.mota(object, numberOfIterations = 50)
r2.mota(object, ...)
mota.progress(value,max.value)
}
\author{Stefan Hengl}

\details{\code{pvalue.mota} and \code{r2.mota} are internal function not of
major interest for the user.}
\keyword{internal}

